python_image_recognistion/requirements.txt
Flask==3.0.0
tensorflow==2.14.0
Pillow==10.0.1
numpy==1.26.0
flask-cors==4